also move the test to tmjob tests
